requirements.txt
annotated-types==0.5.0
blinker==1.6.2
blis==0.7.10
catalogue==2.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
confection==0.1.1
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
dnspython==2.4.1
Flask==2.3.2
Flask-DotEnv==0.1.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
fonttools==4.42.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.3
matplotlib==3.7.2
murmurhash==1.0.9
nltk==3.8.1
numpy==1.25.2
packaging==23.1
pandas==2.0.3
pathy==0.10.2
Pillow==10.0.0
preshed==3.0.8
psycopg2-binary==2.9.7
pydantic==2.1.1
pydantic_core==2.4.0
pymongo==4.4.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.6.3
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
smart-open==6.3.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.19
srsly==2.4.7
thinc==8.1.11
threadpoolctl==3.2.0
tqdm==4.65.1
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wasabi==1.1.2
Werkzeug==2.3.6
WTForms==3.0.1
